Use tsx in date range picker examples
