+++ docs/next.config.js
@@ -42,7 +42,7 @@ module.exports = withBundleAnalyzer(
 
               // Process examples to inject raw code strings
               config.module.rules.push({
-                test: /\.example\.(js|jsx)$/,
+                test: /\.example\.(js|jsx|tsx|ts)$/,
                 include: [path.resolve(__dirname, 'pages')],
                 use: [
                   { loader: 'next-babel-loader' },
