--- docs/pages/demo/daterangepicker/MinMaxDateRangePicker.example.jsx
@@ -1,29 +1,0 @@
-import React, { useState } from 'react';
-import isWeekend from 'date-fns/isWeekend';
-// this guy required only on the docs site to work with dynamic date library
-import { makeJSDateObject } from '../../../utils/helpers';
-import { DateRangePicker, useUtils } from '@material-ui/pickers';
-
-function disableWeekends(date) {
-  // TODO: replace with implementation for your date library
-  return isWeekend(makeJSDateObject(date));
-}
-
-function MinMaxDateRangePicker() {
-  const utils = useUtils();
-  /** @type import('@material-ui/pickers').DateRange */
-  const initialValue = [new Date(), null];
-  const [selectedRange, handleDateChange] = useState(initialValue);
-
-  return (
-    <DateRangePicker
-      disablePast
-      value={selectedRange}
-      shouldDisableDate={disableWeekends}
-      maxDate={utils.addMonths(utils.date(selectedRange[0]), 12)}
-      onChange={date => handleDateChange(date)}
-    />
-  );
-}
-
-export default MinMaxDateRangePicker;
